+++ cpp98/Manual/simpleFilesystem/Client.cpp
@@ -16,7 +16,7 @@ using namespace Filesystem;
 // parameter is the current nesting level (for indentation).
 
 static void
-listRecursive(const DirectoryPrx& dir, int depth = 0)
+listRecursive(const DirectoryPrx& dir, size_t depth = 0)
 {
     string indent(++depth, '\t');
